+++ src/modules/mo-namespace/data-transfer-objects/traffic/compute-time-series-response.dto.ts
@@ -1,0 +1,10 @@
+import { Expose, Transform, Type } from 'class-transformer';
+import { isArray } from 'class-validator';
+import { ComputeTimeSeriesDto } from './compute-time-series.dto';
+
+export class ComputeTimeSeriesResponseDto {
+  @Type(() => ComputeTimeSeriesDto)
+  @Transform(({ value }) => (value && isArray(value) ? value : []))
+  @Expose()
+  results: ComputeTimeSeriesDto[];
+}