+++ src/modules/mo-namespace/data-transfer-objects/traffic/compute-time-series.dto.ts
@@ -1,0 +1,34 @@
+import { Expose, Transform, Type } from 'class-transformer';
+import moment from 'moment';
+import { isNumber } from 'class-validator';
+
+export class ComputeTimeSeriesDto {
+  @Type(() => Number)
+  @Transform(({ value }) => (isNumber(value) ? value : 0))
+  @Expose()
+  cpu: number;
+
+  @Type(() => Number)
+  @Transform(({ value }) => (isNumber(value) ? value : 0))
+  @Expose()
+  cpuLimit: number;
+
+  @Type(() => Number)
+  @Transform(({ value }) => (isNumber(value) ? value : 0))
+  @Expose()
+  memory: number;
+
+  @Type(() => Number)
+  @Transform(({ value }) => (isNumber(value) ? value : 0))
+  @Expose()
+  memoryLimit: number;
+
+  @Type(() => Number)
+  @Transform(({ value }) => (isNumber(value) ? value : 0))
+  @Expose()
+  ephemeralStorageLimit: number;
+
+  @Transform(({ value }) => (value && value !== 'undefined' && value !== 'null' ? moment(value).toDate() : value))
+  @Expose()
+  timeStamp: Date;
+}