+++ src/modules/mo-namespace/data-transfer-objects/traffic/index.ts
@@ -1,1 +1,5 @@
 export * from './origin-traffic.dto';
+export * from './traffic-time-series.dto';
+export * from './traffic-time-series-response.dto';
+export * from './compute-time-series.dto';
+export * from './compute-time-series-response.dto';