+++ src/modules/mo-namespace/data-transfer-objects/traffic/traffic-time-series-response.dto.ts
@@ -1,0 +1,10 @@
+import { Expose, Transform, Type } from 'class-transformer';
+import { TrafficTimeSeriesDto } from './traffic-time-series.dto';
+import { isArray } from 'class-validator';
+
+export class TrafficTimeSeriesResponseDto {
+  @Type(() => TrafficTimeSeriesDto)
+  @Transform(({ value }) => (value && isArray(value) ? value : []))
+  @Expose()
+  results: TrafficTimeSeriesDto[];
+}